+++ src/program/Config.cpp
@@ -132,6 +132,7 @@ void Config::save(const std::string& gamepath) {
 
     settings.setValue("speed_divisor", sc.speed_divisor);
     settings.setValue("fastforward_mode", sc.fastforward_mode);
+    settings.setValue("fastforward_render", sc.fastforward_render);
     settings.setValue("logging_status", sc.logging_status);
     settings.setValue("includeFlags", sc.includeFlags);
     settings.setValue("excludeFlags", sc.excludeFlags);
@@ -277,6 +278,7 @@ void Config::load(const std::string& gamepath) {
 
     sc.speed_divisor = settings.value("speed_divisor", sc.speed_divisor).toInt();
     sc.fastforward_mode = settings.value("fastforward_mode", sc.fastforward_mode).toInt();
+    sc.fastforward_render = settings.value("fastforward_render", sc.fastforward_render).toInt();
     sc.logging_status = settings.value("logging_status", sc.logging_status).toInt();
     sc.includeFlags = settings.value("includeFlags", sc.includeFlags).toInt();
     sc.excludeFlags = settings.value("excludeFlags", sc.excludeFlags).toInt();